Refactor index into metro-transit
